Update Main.py getting two file for config menus and game menus
